requirements.txt
from_root
pandas
numpy
scikit-learn
transformers
python-box==6.0.2
ensure==1.0.2
torch
transformers[torch]

-e .